feat: MeetingResult에서 footer를 분리
