+++ src/pages/MeetingResult/MeetingResult.tsx
@@ -1,34 +1,29 @@
-import Button from '@mui/material/Button';
 import { useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
 import { useRecoilState } from 'recoil';
 
 import { Voting } from '../../apis/votes';
 import { Loading } from '../../components/Loading';
-import { Contents, Footer, Page } from '../../components/pageLayout';
-import { FlexVertical, FullHeightButtonGroup } from '../../components/styled';
+import { Contents, Page } from '../../components/pageLayout';
+import { FlexVertical } from '../../components/styled';
 import { useMeeting } from '../../hooks/useMeeting';
-import useShare from '../../hooks/useShare';
 import { useVotings } from '../../hooks/useVotings';
 import { votingsState } from '../../stores/voting';
 import MeetingResultContents from '../../templates/MeetingResult/MeetingResultContents';
+import MeetingResultFooter from '../../templates/MeetingResult/MeetingResultFooter';
 import MeetingResultHeader from '../../templates/MeetingResult/MeetingResultHeader';
 
 function MeetingResult() {
-  const navigate = useNavigate();
   const [, setVotings] = useRecoilState<Voting[]>(votingsState);
-  const { openShare, setTarget } = useShare();
-  const { meeting, meetingId, isFetching: isMeetingFetching } = useMeeting();
-  const { votings, isFetching: isVotingsFetcing } = useVotings();
+  const { isFetching: isMeetingFetching } = useMeeting();
+  const { votings, isLoading, isFetching: isVotingsFetcing } = useVotings();
   const isFetching = isMeetingFetching && isVotingsFetcing;
 
   useEffect(() => {
-    if (meeting && votings) {
+    if (!isLoading && votings) {
       setVotings(votings);
-      setTarget(meeting);
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [meeting, votings, meetingId, setVotings]);
+  }, [isLoading]);
 
   if (isFetching) {
     return <Loading />;
@@ -42,31 +37,7 @@ function MeetingResult() {
           <MeetingResultContents />
         </FlexVertical>
       </Contents>
-
-      <Footer>
-        <FullHeightButtonGroup
-          fullWidth
-          disableElevation
-          variant="contained"
-          aria-label="Disabled elevation buttons"
-        >
-          <Button
-            color="secondary"
-            onClick={() => {
-              meeting?.id && navigate(`/meetings/${meeting?.id}`);
-            }}
-          >
-            상세보기
-          </Button>
-          <Button
-            onClick={() => {
-              openShare();
-            }}
-          >
-            공유
-          </Button>
-        </FullHeightButtonGroup>
-      </Footer>
+      <MeetingResultFooter />
     </Page>
   );
 }
